sf methods for post_table objects

sf methods for post_table objects

Usage

# S3 method for class 'post_table'
st_as_sf(x, ...)

# S3 method for class 'post_table'
st_crs(x) <- value

st_set_crs.post_table(x, value)

# S3 method for class 'post_table'
st_normalize(x, domain, ...)

# S3 method for class 'post_table'
st_shift_longitude(x, ...)

# S3 method for class 'post_table'
st_transform(x, crs, ...)

# S3 method for class 'post_table'
st_wrap_dateline(x, ...)

# S3 method for class 'post_table'
st_zm(x, ...)

Arguments

x

a post_table object

...

additional arguments passed onto the respective sf function

value

one of (i) character: a string accepted by GDAL, (ii) integer, a valid EPSG value (numeric), or (iii) an object of class crs.

domain

The domain x should be normalized from as a length 4 vector of the form c(xmin, ymin, xmax, ymax). Defaults to the bounding box of x

crs

target coordinate reference system: object of class crs, or input string for st_crs

Details

sf::st_as_sf() for post_table objects sets the changing geometry as the active sf_column but keeps the summarised geometry in a separate column.

st_normalize takes as domain the bounding box of the post_table in the nested form (spatial face)