Svalbard glaciers subset to exemplify post functions

Selected glacier from the digital glacier dataset compiled and released in König et al. (2013).

Format

An object of class sf with POLYGON geometries with 25 rows and 3 columns:

name

Glacier name

year

The year of glacier delineation

length

Maximum glacier length from the glacier tongue to the head of the accumulation area

fwidth

Average width of the glacier tongue

geom

the geometry list column

Details

When the glaciers share the same name for a single year, the polygons are combined into a MULTIPOLYGON. In these cases, the maximum LENGTH and FWIDTH are taken for the feature.

The two detached glaciers from Midtre Lovenbreen for the year 2007 where delineated originally in 2009. The year was changed to 2007 to keep congruent data.

References

König, M., Kohler J., and Nuth C. (2013). “Glacier Area Outlines - Svalbard 1936-2010.” Norwegian Polar Institute. https://doi.org/10.21334/npolar.2013.89f430f8.

König, M., Nuth, C., Kohler, J., Moholdt, G. and Pettersen, R. (2014). “A Digital Glacier Database for Svalbard.” Global Land Ice Measurements from Space, 229–39. Springer Berlin Heidelberg. https://doi.org/10.1007/978-3-540-79818-7_10.

Examples

library(sf)
svalbard = st_read(system.file("extdata/svalbard.gpkg", package="post"))
#> Reading layer `svalbard_glaciers_subset' from data source 
#>   `/home/runner/work/_temp/Library/post/extdata/svalbard.gpkg' 
#>   using driver `GPKG'
#> Simple feature collection with 15 features and 4 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 428910.1 ymin: 8754278 xmax: 440637.1 ymax: 8762561
#> Projected CRS: WGS 84 / UTM zone 33N