Summarise changing geometries based on a group identifier

Changing geometries are summarised for all time periods as a single geometry (of types POINT, POLYGON, MULTIPOLYGON) in order to pass them as a spatial dimension to post_* objects.

summarise_geometry_union computes the geometry summary as the union and dissolve of the changing geometries

summarise_geometry_centroid computes the geometry summary as the centroid of the union and dissolve of the changing geometries

summarise_geometry_bbox computes the geometry summary as the bounding box of the union and dissolve of the changing geometries

Usage

summarise_geometry_union(x, group_id = NULL, sf_column_name = NULL)

summarize_geometry_union(x, group_id = NULL, sf_column_name = NULL)

summarise_geometry_centroid(x, group_id = NULL, sf_column_name = NULL)

summarize_geometry_centroid(x, group_id = NULL, sf_column_name = NULL)

summarise_geometry_bbox(x, group_id = NULL, sf_column_name = NULL)

summarize_geometry_bbox(x, group_id = NULL, sf_column_name = NULL)

Arguments

x

object POLYGON/MULTIPOLYGON changing geometries to summarise based on group_id.

group_id

(character) name of column containing the grouping identifier for the changing polygons or group identifier given as an integer or character value that is assigned to all rows in x. Defaults to the first column in x.

sf_column_name

(character) name of column with changing geometries. Defaults to active sf_column.

Value

object of class sfc with summarised geometries based on group_id

Details

The family of functions summarise_geometry_*() or summarize_geometry_*() are helpers to create these summarised geometries. Supported summary types are:

a. the union and dissolve of the changing geometries,

b. the centroid of a

c. the bbox of a

A custom function to summarise geometries can be passed onto the coercion functions as_post_*(), given that the parameters x, group_id and sf_column_name are included in the function. Additional arguments can be passed to the function as necessary.

Examples

summarise_geometry_union(polygons)
#> Geometry set for 5 features 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -0.2974337 ymin: -0.0029756 xmax: 0.9730806 ymax: 1.153558
#> Geodetic CRS:  WGS 84
#> POLYGON ((0.4971355 0.8717961, 0.5474949 0.8088...
#> POLYGON ((-0.1774445 0.5995572, -0.1843257 0.58...
#> POLYGON ((0.28657 0.6085207, 0.2663408 0.622264...
#> POLYGON ((0.5737123 0.4613799, 0.6063966 0.4263...
#> POLYGON ((0.2973931 0.2557416, 0.2813385 0.2544...
summarise_geometry_centroid(polygons)
#> Geometry set for 5 features 
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -0.002543201 ymin: 0.17772 xmax: 0.752585 ymax: 0.9018588
#> Geodetic CRS:  WGS 84
#> POINT (0.647816 0.9018588)
#> POINT (-0.002543201 0.8403816)
#> POINT (0.3897167 0.5412363)
#> POINT (0.752585 0.3538271)
#> POINT (0.4690683 0.17772)
summarise_geometry_bbox(polygons)
#> Geometry set for 5 features 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -0.2974337 ymin: -0.0029756 xmax: 0.9730806 ymax: 1.153558
#> Geodetic CRS:  WGS 84
#> POLYGON ((0.4361026 0.7456421, 0.9005707 0.7456...
#> POLYGON ((-0.2974337 0.5814869, 0.3796448 0.581...
#> POLYGON ((0.2145846 0.4039274, 0.5898739 0.4039...
#> POLYGON ((0.5272362 0.1748722, 0.9730806 0.1748...
#> POLYGON ((0.2452129 -0.0029756, 0.6988282 -0.00...